pyproject.toml
[tool.poetry]
name = "chess"
version = "0.1.0"
description = "Chess"

[tool.poetry.dependencies]
python = "^3.10"
cadbuildr-foundation = "^0.0.0.13"

[tool.cadbuildr]
logo = "logo.png"

[tool.cadbuildr.dependencies]
cadbuildr-foundation = "^0.0.0.13"
